deployer: improve validation event handling
Updated validations to look at Scheduled events as an antidote to
SchedulingFailed events and give the failures a second chance to be
resolved. This should address the case where a PVC is dynamically bound,
the pod having received a SchedulingFailed in the meantime.

Also, modify deployer pod definition to prevent it being put in the
anyuid scc where it couldn't change its own labels.