[BZ 1473013] disable commit log compression
We have been using commit log compression, but other commit log
settings in cassandra.yaml assume compression is disabled. I think
this is why we have run into problems with running out of memory
during commit log replay.

Compression was turned on as an optimization. The safest course
at least for now is to simply revert back to out of the box
settings.