Add support for building dev images
To date there has been no support for building images with locally built
versions of hawkular-metrics. The docker build only produces images that
include released versions of hawkular-metrics that get downloaded from
the JBoss nexus repo. This makes it awkward and error prone to test
changes that have not been released. I have resorted to modifying the
Dockerfile to copy a locally built hawkular-metrics EAR.

This commit adds support for building images with locally built versions
with a few small changes. The build-images.sh script has a --dev option
to indicate a dev build. If the dev flag is set and if a
hawkular-metrics.ear file exists in the hawkular-metrics/dev directory,
then it will be used. The hawkular-metrics version environment variable
will be ignored.