[BZ 1567251] Disable auto snapshots and env var for taking snapshots
Disabling auto snapshots for Cassandra by default because 1) we do not
have a mechanism for snapshots and because 2) the temp data temp tables
introduced in OpenShift 3.7 generate a lot of snapshots that can eat up
disk space in a hurry.

This commit introduces a new environment variable, TAKE_SNAPSHOT, which
if set to "true" will cause a snapshot to be generated in the
cassandra-poststart.sh script which is run from the postStart lifecycle
hook. The snapshop name uses a default value of a timestamp.